pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ldraney-linkedin-mcp"
version = "0.1.0"
description = "MCP server wrapping the LinkedIn Python SDK (v202510)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0",
    "ldraney-linkedin-sdk>=0.1.0",
    "keyring>=25.0",
]

[project.scripts]
ldraney-linkedin-mcp = "linkedin_mcp.server:main"
linkedin-mcp-scheduler = "linkedin_mcp.scheduler_db:run_scheduler"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
ldraney-linkedin-sdk = { path = "../linkedin-sdk" }

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
